feat: cancel subsequent bookigns
