+++ packages/emails/templates/organizer-cancelled-email.ts
@@ -1,4 +1,5 @@
 import { EMAIL_FROM_NAME } from "@calcom/lib/constants";
+import type { CalendarEvent, Person } from "@calcom/types/Calendar";
 
 import { renderEmail } from "../";
 import generateIcsFile, { GenerateIcsRole } from "../lib/generateIcsFile";
@@ -20,11 +21,15 @@ export default class OrganizerCancelledEmail extends OrganizerScheduledEmail {
         title: this.calEvent.title,
         date: this.getFormattedDate(),
       })}`,
-      html: await renderEmail("OrganizerCancelledEmail", {
-        attendee: this.calEvent.organizer,
-        calEvent: this.calEvent,
-      }),
+      html: await this.getHtml(this.calEvent, this.calEvent.organizer),
       text: this.getTextBody("event_request_cancelled"),
     };
+  }
+
+  protected async getHtml(calEvent: CalendarEvent, organizer: Person) {
+    return await renderEmail("OrganizerCancelledEmail", {
+      calEvent,
+      attendee: organizer,
+    });
   }
 }
