+++ packages/platform/libraries/index.ts
@@ -5,8 +5,10 @@ import { CalendarService as IcsFeedCalendarService } from "@calcom/app-store/ics
 import type { CredentialOwner } from "@calcom/app-store/types";
 import { getAppFromSlug } from "@calcom/app-store/utils";
 import type { CredentialDataWithTeamName, LocationOption } from "@calcom/app-store/utils";
+import AttendeeCancelledEmail from "@calcom/emails/templates/attendee-cancelled-email";
 import AttendeeDeclinedEmail from "@calcom/emails/templates/attendee-declined-email";
 import AttendeeScheduledEmail from "@calcom/emails/templates/attendee-scheduled-email";
+import OrganizerCancelledEmail from "@calcom/emails/templates/organizer-cancelled-email";
 import OrganizerScheduledEmail from "@calcom/emails/templates/organizer-scheduled-email";
 import { getBookingForReschedule } from "@calcom/features/bookings/lib/get-booking";
 import getBookingInfo from "@calcom/features/bookings/lib/getBookingInfo";
@@ -194,3 +196,7 @@ export { AttendeeScheduledEmail };
 export { OrganizerScheduledEmail };
 
 export { AttendeeDeclinedEmail };
+
+export { AttendeeCancelledEmail };
+
+export { OrganizerCancelledEmail };
